+++ NEWS.md
@@ -1,10 +1,17 @@
 # labelled (development version)
 
+* new functions `user_na_to_tagged_na()` and `tagged_na_to_user_na()`
+* new option `explicit_tagged_na` in `to_factor()` and `to_character()`
+* new function `is_user_na()`
 * trying to apply a value label, `na_range()` or `na_values()` to a factor 
   will now produce an error
 * new set of unit tests (#99)
+* new functions `user_na_to_tagged_na()` and `tagged_na_to_user_na()`
+* new option `explicit_tagged_na` in `to_factor()` and `to_character()`
+* new function `is_user_na()`
 * new function `look_for_and_select()` (#87)
 * `look_for()` can now search within factor levels and value labels (#104)
+* better printing of value labels (#89)
 * bug fix in `foreign_to_labelled()` for Stata files (#100)
 
 # labelled 2.8.0